+++ docs/changes.rst
@@ -15,6 +15,8 @@ Change Log
   * created :meth:`.get_d` to copy the remote directory to a local path (non-recursive)
   * added :attr:`.timeout` to set the read/write timeout of the underlying channel for pending read/write ops
   * added :meth:`.listdir_attr`, wrapper for paramiko method
+  * added :meth:`.truncate`, method returns the new file size
+
 
 * 0.2.6 (current, released 2014-05-17)